# Optimize find_yregs pass from beam_ssa_pre_codegen