The current pass gets very expensive the longer the
code block is. This commit address it. For example,
imagine this code:

    -module(bar).
    -export([bar/0]).
    bar() ->
      foo:bar(),
      % repeat foo:bar() 999 times
      ok.

Prior to this patch, it would take:

    find_yregs                 :      0.212 s  73 %

After this patch, it takes:

    find_yregs                 :      0.013 s  14 %

Another example is the code in this gist:
https://gist.github.com/josevalim/acac154996e1e61058f367a5b3a161aa

After the patch, find_yregs goes from 3.5s to 0.4s.

This patch is not ready yet for merge, please see comments inline for
places where I would like style-related feedback.