@josevalim 2020-09-17
Finally, I would like to mention that this patch would be simpler if we implemented [EEP 50](https://www.erlang.org/erlang-enhancement-proposals/eep-0050.html). Today, I have to change data structures and the algorithm at the same time, which complicates things. With EEP 50, I would only have to worry about the algorithm.

The other aspect to note is that this optimization could be applied elsewhere, for example in `beam_ssa:def_unused/3` but we can't do so easily because `def_unused/3` works with a different set than the cerl_sets used in this PR (ordsets vs cerl_sets). With EEP 50, it would be easier to reuse those algorithms too.