@josevalim 2020-09-18
Code simplified! This should be good to go now!